requirements.txt
transformers
pandas
torch
numpy
matplotlib
seaborn
nbformat
pyyaml
openai
